fix (#78): erro de tipagem

+++ src/main/java/br/com/chronos/servicologin/autenticacao/PayloadUsuario.java
@@ -3,10 +3,10 @@ package br.com.chronos.servicologin.autenticacao;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record PayloadUsuario(
-        Long id,
+        Integer id,
         String nome,
         String email,
         @JsonProperty("cargo_id")
-        Integer cargoId
+        Long cargoId
 ) {
 }
