+++ src/main/java/br/com/chronos/servicologin/usuario/Usuario.java
@@ -1,10 +1,10 @@
 package br.com.chronos.servicologin.usuario;
 
 public record Usuario(
-        Long id,
+        Integer id,
         String nome,
         String email,
         String passwordHash,
-        Integer cargoId
+        Long cargoId
 ) {
 }
